moving ui state files to specific package
