+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/BalanceCard.kt
@@ -55,7 +55,6 @@ import br.com.rbrthmn.ui.financialcompanion.components.AddItemButton
 import br.com.rbrthmn.ui.financialcompanion.components.BanksDropdownMenu
 import br.com.rbrthmn.ui.financialcompanion.components.DecimalInputField
 import br.com.rbrthmn.ui.financialcompanion.components.TotalValueText
-import br.com.rbrthmn.ui.financialcompanion.uistates.BankAccountBalanceUiState
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
