+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/BalanceCardUiState.kt
@@ -18,15 +18,9 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.uistates
+package br.com.rbrthmn.ui.financialcompanion.screens.home.components
 
-import br.com.rbrthmn.R
-
-data class CreditCardBillUiState(
-    val name: String = "",
-    val value: String = "",
-    val bankName: String = "",
-    val bankIcon: Int = R.drawable.bank_icon,
-    val canValueBeEdited: Boolean = false,
-    val dueDay: Int = 0
+data class BalanceCardUiState(
+    val totalBalance: String = "",
+    val accounts: List<BankAccountBalanceUiState> = listOf()
 )
